requirements.txt
pymongo
python-dotenv
ratelimiter
mysql-connector-python
requests
